% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_alias.R
\name{check_crucial_names}
\alias{check_crucial_names}
\title{Check if a named object contains expected names}
\usage{
check_crucial_names(x, expected_names)
}
\arguments{
\item{x}{A named object.}

\item{expected_names}{String; expected names of \code{x}.}
}
\value{
Invisible \code{x}, or an error with informative message.

Adapted from: https://github.com/RMI-PACTA/r2dii.match/blob/main/R/check_crucial_names.R
}
\description{
Based on fgeo.tool::check_crucial_names()
}
\examples{
x <- c(a = 1)
check_crucial_names(x, "a")
try(check_crucial_names(x, "bad"))
}
