% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_duplicated_score}
\alias{check_duplicated_score}
\title{Reports duplicates id on manual matched errors}
\usage{
check_duplicated_score(manually_matched)
}
\arguments{
\item{manually_matched}{Tibble holding the result of the matching process, after the
user has manually verified and matched the results}
}
\value{
Input \code{manually_matched}
}
\description{
Reports duplicates id on manual matched errors
}
