% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_match.R
\name{check_loanbook}
\alias{check_loanbook}
\title{Checks your \code{loanbook} is as we expect}
\usage{
check_loanbook(loanbook)
}
\arguments{
\item{loanbook}{Your \code{loanbook} dataset.}
}
\value{
Called for it's side effects. Returns \code{loanbook} invisibly.
}
\description{
Checks your \code{loanbook} is as we expect
}
\examples{
library(vroom)

loanbook <- vroom(example_file("demo_loanbook.csv"), show_col_types = FALSE)
check_loanbook(loanbook)
}
