% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demo_loanbook.R
\docType{data}
\name{demo_loanbook}
\alias{demo_loanbook}
\title{Demo loanbook entries}
\format{
A tibble with 11 rows and 5 variables:
\describe{
\item{id}{a numeric id}
\item{company_name}{name of company}
\item{postcode}{postcode of company}
\item{country}{country name in lowercase}
\item{misc_info}{A placeholder column that holds additional information that human matchers would consider in matching}
}
}
\usage{
demo_loanbook
}
\description{
A simplified demo tilt company data set to illustrate and test matching with
loanbook. For details on included cases please refer to data generation
script.
}
\examples{
demo_loanbook
}
\keyword{datasets}
