% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demo_matched.R
\docType{data}
\name{demo_matched}
\alias{demo_matched}
\title{Demo matched db entries}
\format{
A tibble with 9 rows and 5 variables:
\describe{
\item{id}{a numeric id}
\item{company_name}{name of company}
\item{postcode}{postcode of company}
\item{country}{country name in lowercase}
\item{misc_info}{A placeholder column that holds additional information that human matchers would consider in matching}
\item{company_alias}{name of the company, preprocessed}
\item{id_tilt}{a numeric id in the tilt db}
\item{company_name_tilt}{name of company}
\item{misc_info_tilt}{A placeholder column that holds additional information that human matchers would consider in matching}
\item{company_alias_tilt}{name of the company, preprocessed}
\item{string_sim}{string similarity between aliased company name in the loanbook and aliase company name in tilt db}
\item{suggest_match}{set to TRUE if string_sim is above a certain threshold}
\item{accept_match}{manual decision to whether the company from the loanbook matches a comapny in the tilt db}
}
}
\usage{
demo_matched
}
\description{
A simplified demo matched company data set to illustrate the manual matching process with
loanbook. For details on included cases please refer to data generation
script.
}
\examples{
demo_matched
}
\keyword{datasets}
