% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_paste.R
\name{duplicated_paste}
\alias{duplicated_paste}
\title{Detect duplicated strings}
\usage{
duplicated_paste(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[base:paste]{base::paste}}
  \describe{
    \item{\code{sep}}{a character string to separate the terms.  Not
    \code{\link[base]{NA_character_}}.}
    \item{\code{collapse}}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.}
    \item{\code{recycle0}}{\code{\link[base]{logical}} indicating if zero-length
    character arguments should lead to the zero-length
    \code{\link[base]{character}(0)} after the \code{sep}-phase (which turns into
    \code{""} in the \code{collapse}-phase, i.e., when \code{collapse} is
    not \code{NULL}).}
  }}
}
\value{
A logical vector of the same length as the longest vector passed to
\code{...}.
}
\description{
Detect duplicated strings
}
\examples{
duplicated_paste(c("a", "a"), 1:2)
paste(c("a", "a"), 1:2)

duplicated_paste(c("a", "a"), c(1, 1))
paste(c("a", "a"), c(1, 1))
}
