% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_file.R
\name{example_file}
\alias{example_file}
\title{Get the path to an example file}
\usage{
example_file(file)
}
\arguments{
\item{file}{Name of the file.}
}
\value{
A path.
}
\description{
Get the path to an example file
}
\examples{
example_file("demo_loanbook.csv")

example_file("demo_tilt.csv")

example_file("demo_matched.csv")
}
