% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_name_to_alias}
\alias{from_name_to_alias}
\title{From name to alias}
\usage{
from_name_to_alias()
}
\value{
\link[tibble:tibble]{tibble::tibble} with columns \code{from} and
\code{to}.
}
\description{
Function that outputs a table giving default strings used to
convert from a name to its alias. You may amend this table and pass it to
\code{to_alias()} via the \code{from_to} argument.
}
\details{
Source: @jdhoffa https://github.com/RMI-PACTA/r2dii.dataraw/pull/8
}
