% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_duplicates}
\alias{report_duplicates}
\title{Report duplicate rows}
\usage{
report_duplicates(data, cols)
}
\arguments{
\item{data}{Tibble holding a result data set.}

\item{cols}{Vector of columns names on which we want to test if there are
duplicates on.}
}
\description{
Reports duplicates in \code{data} on columns \code{cols}. More specifically, we are
interested in this case on the \code{company_name}, \code{postcode} and \code{country}
columns. Duplicates are reported via a warning.
}
\keyword{internal}
