% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_missings}
\alias{report_missings}
\title{Report missing}
\usage{
report_missings(data)
}
\arguments{
\item{data}{Tibble holding a result data set.}
}
\value{
Input \code{data}.
}
\description{
Function that reports number of missing values per columns found in the data set.
}
