% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_no_matches.R
\name{report_no_matches}
\alias{report_no_matches}
\title{Reports companies that were not matched in the loanbook}
\usage{
report_no_matches(loanbook, manually_matched)
}
\arguments{
\item{loanbook}{Loanbook data set}

\item{manually_matched}{Tibble holding the result of the matching process,
after the user has manually selected and matched the companies in the
loanbook with the tilt data set.}
}
\value{
\code{not_matched_companies} Tibble holding id and company name of the
companies not matched by the tilt data set.
}
\description{
Reports companies that were not matched in the loanbook
}
\examples{
library(tibble)

loanbook <- tibble(id = 1:2, company_name = letters[id], irrelevant = "xyz")
loanbook

accepted <- tibble(id = 1:2, accept_match = c(TRUE, FALSE))
accepted

report_no_matches(loanbook, accepted)

# It's rigurous but fails with informative messages:
# The names of crucial columns must be as documented.
try(report_no_matches(loanbook, tibble(ids = 1, accept_match = TRUE)))

# The type of `accept_match` must be as documented.
try(report_no_matches(loanbook, tibble(id = 1, accept_match = "TRUE")))
}
