% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_match.R
\name{suggest_match}
\alias{suggest_match}
\title{Given a loanbook and tilt datasets returns a dataset with suggested matches}
\usage{
suggest_match(
  loanbook,
  tilt,
  eligibility_threshold = 0.75,
  suggestion_threshold = 0.9
)
}
\arguments{
\item{loanbook}{Path to a .csv file with your \code{loanbook} data.}

\item{tilt}{Path to a .csv file with 2DII's \code{tilt} data.}

\item{eligibility_threshold}{Eligibility threshold.}

\item{suggestion_threshold}{Suggestion threshold.}
}
\value{
A dataframe with suggested matching candidates.
}
\description{
Given a loanbook and tilt datasets returns a dataset with suggested matches
}
\examples{
library(vroom)
loanbook <- vroom(example_file("demo_loanbook.csv"), show_col_types = FALSE)
tilt <- vroom(example_file("demo_tilt.csv"), show_col_types = FALSE)

suggest_match(loanbook, tilt)
}
