% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt.company.match-package.R
\docType{package}
\name{tilt.company.match-package}
\alias{tilt.company.match}
\alias{tilt.company.match-package}
\title{tilt.company.match: Helpers for Company Name Matching in the tilt-Project}
\description{
Package holds helpers for company matching in the tilt-project. For details please refer to README.
}
\author{
\strong{Maintainer}: Mirja Hollmann \email{mirja@2degrees-investing.org}

Authors:
\itemize{
  \item Linda Delacombaz \email{linda@2degrees-investing.org}
}

}
\keyword{internal}
