% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_paste.R
\name{duplicated_paste}
\alias{duplicated_paste}
\title{Duplicated paste}
\usage{
duplicated_paste(...)
}
\arguments{
\item{...}{Typically columns of a data set, e.g. company_name and postcode.}
}
\value{
A vector of booleans: TRUE if the value is a duplicate, FALSE if not.
}
\description{
\code{duplicated_paste} check for duplicates on multiple rows / columns.
}
