% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_paste.R
\name{duplicated_paste}
\alias{duplicated_paste}
\title{Detect duplicated strings}
\usage{
duplicated_paste(...)
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}
}
\description{
Detect duplicated strings
}
\examples{
duplicated_paste(c("a", "a"), 1:2)
paste(c("a", "a"), 1:2)

duplicated_paste(c("a", "a"), c(1, 1))
paste(c("a", "a"), c(1, 1))
}
