% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_unmatched.R
\name{pick_unmatched}
\alias{pick_unmatched}
\title{Pick unmatched \code{id} and \code{company_name} from a loanbook}
\usage{
pick_unmatched(loanbook, accepted)
}
\arguments{
\item{loanbook}{A data frame with columns \code{id} and \code{company_name}.}

\item{accepted}{A data frame with columns \code{id} and \code{accept_match} (of type
logical).}
}
\value{
A data frame with columns \code{id} and \code{company_name} reflecting the
\code{looanbook} rows that don't match the \code{accepted} data frame.
}
\description{
Pick unmatched \code{id} and \code{company_name} from a loanbook
}
\examples{
library(tibble)

loanbook <- tibble(id = 1:2, company_name = letters[id], irrelevant = "xyz")
loanbook

accepted <- tibble(id = 1:2, accept_match = c(TRUE, FALSE))
accepted

pick_unmatched(loanbook, accepted)

# It's rigurous but fails with informative messages:
# The names of crucial columns must be as documented.
try(pick_unmatched(loanbook, tibble(ids = 1, accept_match = TRUE)))

# The type of `accept_match` must be as documented.
try(pick_unmatched(loanbook, tibble(id = 1, accept_match = "TRUE")))
}
